refactor: extract RecordingSessionCancelStatusPresenter.swift from RecordingSessionPresenters.swift
Splits the `RecordingSessionCancelStatusPresenter` (~32 lines) out of
RecordingSessionPresenters.swift into its own file.

Byte-identical move of lines 85-116. Both files import Combine +
Foundation.

RecordingSessionPresenters.swift: 190 → 156 lines. Tests: 667.
